requirements.txt
streamlit
python-dotenv
google-generativeai
requests
pandas
numpy